feat: move ensureCommandShape into common (required for neovim)
