+++ packages/common/src/index.ts
@@ -94,9 +94,11 @@ export * from "./util/itertools";
 export * from "./extensionDependencies";
 export * from "./FakeCommandServerApi";
 export * from "./types/TestCaseFixture";
+export * from "./util/ensureCommandShape";
 export * from "./util/getEnvironmentVariableStrict";
 export * from "./util/CompositeKeyDefaultMap";
 export * from "./util/toPlainObject";
+export * from "./util/type";
 export * from "./util/clientSupportsFallback";
 export * from "./scopeSupportFacets/scopeSupportFacets.types";
 export * from "./scopeSupportFacets/scopeSupportFacetInfos";
