+++ packages/cursorless-engine/src/cursorlessEngine.ts
@@ -4,6 +4,7 @@ import {
   FileSystem,
   Hats,
   IDE,
+  ensureCommandShape,
   ScopeProvider,
 } from "@cursorless/common";
 import {
@@ -14,7 +15,6 @@ import { Debug } from "./core/Debug";
 import { HatTokenMapImpl } from "./core/HatTokenMapImpl";
 import { Snippets } from "./core/Snippets";
 import { StoredTargetMap } from "./core/StoredTargets";
-import { ensureCommandShape } from "./core/commandVersionUpgrades/ensureCommandShape";
 import { RangeUpdater } from "./core/updateSelections/RangeUpdater";
 import { CustomSpokenFormGeneratorImpl } from "./generateSpokenForm/CustomSpokenFormGeneratorImpl";
 import { LanguageDefinitions } from "./languages/LanguageDefinitions";
